rename page components to comply with naming scheme
